pyproject.toml
[build-system]
requires = ["setuptools>=70.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "loom"
version = "0.0.1"
description = "Declarative orchestration framework for AI pipelines with built-in evaluation"
authors = [{name = "Evan Volgas", email = "evan.volgas@gmail.com"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10,<3.14"
keywords = ["ai", "llm", "pipeline", "orchestration", "evaluation", "etl", "aietl"]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

# Core dependencies - minimal for basic functionality
dependencies = [
    "arbiter @ git+https://github.com/evanvolgas/arbiter.git",  # Hard dependency for evaluation
    "pydantic>=2.12.0",
    "pyyaml>=6.0.0",
    "polars>=0.19.0",  # DataFrame processing (faster than pandas)
    "pyarrow>=14.0.0",  # Parquet support
    "click>=8.0.0",  # CLI framework
    "httpx>=0.28.0",  # Async HTTP (for future HTTP sources)
    "python-dotenv>=1.0.1",
    "aiofiles>=25.0.0",
    "rich>=14.0.0",  # CLI formatting and progress bars
    "typing-extensions>=4.0.0",
]

[project.optional-dependencies]
# Database connectors (Phase 2+)
postgres = [
    "psycopg[binary,pool]>=3.0.0",
    "sqlalchemy>=2.0.0",
]

# Cache and storage backends (Phase 3+)
redis = ["redis>=5.0.0"]

# Streaming support (Phase 7)
streaming = ["bytewax>=0.20.0"]

# Development tools
dev = [
    # Testing
    "pytest>=9.0.0",
    "pytest-cov>=7.0.0",
    "pytest-asyncio>=1.0.0",
    "pytest-mock>=3.15.0",
    "pytest-xdist>=3.8.0",  # Parallel test execution

    # Code quality
    "black>=25.0.0",
    "mypy>=1.18.0",
    "ruff>=0.14.0",
    "types-pyyaml>=6.0.0",
    "types-aiofiles>=25.0.0",

    # Interactive development
    "ipython>=9.0.0",
    "ipdb>=0.13.13",

    # Documentation
    "mkdocs>=1.6.0",
    "mkdocs-material>=9.7.0",
    "mkdocstrings[python]>=0.30.0",

    # Development utilities
    "pre-commit>=4.0.0",
    "watchdog>=6.0.0",
]

# All features
all = ["loom[postgres,redis,streaming,dev]"]

[project.scripts]
loom = "loom.cli.main:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["loom*"]

[tool.setuptools.package-data]
loom = ["py.typed"]

[tool.black]
line-length = 88
target-version = ['py310', 'py311', 'py312', 'py313']

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
]

ignore = [
    "E501",   # line too long (handled by formatter)
]

fixable = ["ALL"]
unfixable = []

[tool.ruff.lint.isort]
known-first-party = ["loom"]
combine-as-imports = true

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.pytest.ini_options]
minversion = "8.0"
addopts = ["--strict-markers", "--cov=loom", "--cov-report=term-missing"]
testpaths = ["tests"]
markers = [
    "slow: marks tests as slow",
    "integration: marks tests as integration tests",
    "performance: marks tests as performance tests",
    "e2e: marks tests as end-to-end pipeline tests",
]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
show_error_codes = true
pretty = true
disallow_any_generics = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_incomplete_defs = true
disallow_untyped_decorators = true
no_implicit_reexport = true
strict_optional = true
strict_equality = true
warn_no_return = true
warn_unreachable = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[[tool.mypy.overrides]]
module = ["polars", "polars.*", "pyarrow", "pyarrow.*"]
ignore_missing_imports = true
